+++ app/models/flow.rb
@@ -124,4 +124,4 @@ class Flow
     FLOWS_MAP[flow_type.to_sym] || []
   end
 end
-# rubocop:enabled ClassLength
+# rubocop:enable ClassLength