+++ app/models/service.rb
@@ -49,6 +49,7 @@ class Service < ApplicationRecord
 
   def check_override(uri)
     return false if uri.blank?
+
     sc = Service.where('uri LIKE ?', "#{uri}%").where.not(id: id)
 
     # Regexp used to eliminate false-positives for TLDs (e.g. .co vs .com)
@@ -57,11 +58,13 @@ class Service < ApplicationRecord
 
   def check_if_not_override_uri
     return unless check_override(uri)
+
     errors.add(:uri, I18n.t('activerecord.errors.models.service.uri.override'))
   end
 
   def check_if_not_overridden_by_uri
     return unless uri.present? && any_other? { |s| uri.start_with? "#{s.uri}/" }
+
     errors.add(:uri, I18n.t('activerecord.errors.models.service.uri.overridden'))
   end
 
@@ -71,12 +74,14 @@ class Service < ApplicationRecord
 
   def check_if_not_overridden_by_alias
     return unless uri.present? && any_other? { |s| check_overridden(uri_aliases, s.uri) }
+
     errors.add(:uri_aliases,
                I18n.t('activerecord.errors.models.service.uri_aliases.overridden'))
   end
 
   def check_if_alias_not_overridden
     return unless uri.present? && any_other? { |s| check_overridden(s.uri_aliases, uri) }
+
     errors.add(:uri_aliases,
                I18n.t('activerecord.errors.models.service.uri_aliases.overridden'))
   end
@@ -85,25 +90,27 @@ class Service < ApplicationRecord
     Service.where.not(id: id).any? { |s| yield(s) }
   end
 
-  def duplicate_aliases?(la)
+  def duplicate_aliases?(uri_alias)
     sql = <<~SQL
       EXISTS (SELECT * FROM (SELECT unnest(services.uri_aliases))
        x(uri_aliases) WHERE x.uri_aliases LIKE ?)
     SQL
-    Service.where(sql, "#{la}%").where.not(id: id).count.positive?
+    Service.where(sql, "#{uri_alias}%").where.not(id: id).count.positive?
   end
 
-  def duplicate_uri?(u)
-    Service.where('uri LIKE ?', "#{u}%").where.not(id: id).count.positive?
+  def duplicate_uri?(uri)
+    Service.where('uri LIKE ?', "#{uri}%").where.not(id: id).count.positive?
   end
 
   def check_uri_aliases_format
     return unless uri_aliases.any? { |u| u !~ /\A#{URI.parser.make_regexp}\z/ }
+
     errors.add(:uri_aliases, I18n.t('activerecord.errors.models.service.uri_aliases.format'))
   end
 
   def check_uri_aliases_same
     return unless uri_aliases.any? { |u| u == uri }
+
     errors.add(:uri_aliases,
                I18n.t('activerecord.errors.models.service.uri_aliases.urialiassame'))
   end
@@ -111,6 +118,7 @@ class Service < ApplicationRecord
   # Checks if any of the aliases overlaps with any of the services' aliases
   def check_uri_aliases_uniqueness
     return unless uri_aliases.any? { |u| duplicate_aliases?(u) }
+
     errors.add(:uri_aliases, I18n.t('activerecord.errors.models.service.uri_aliases.uniqueness'))
   end
 end