+++ app/services/pipelines/start_runnable.rb
@@ -19,7 +19,20 @@ module Pipelines
 
     def runnable?(computation)
       computation.runnable? &&
-        (!computation.rimrock? || @user_proxy.valid?)
+        valid_proxy?(computation) &&
+        configured?(computation)
+    end
+
+    def valid_proxy?(computation)
+      !computation.rimrock? || @user_proxy.valid?
+    end
+
+    def configured?(computation)
+      if computation.rimrock?
+        computation.tag_or_branch.present?
+      elsif computation.webdav?
+        computation.run_mode.present?
+      end
     end
   end
 end